# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  controllers.MaterialCreateRequest:
    properties:
      title_cn:
        type: string
      title_en:
        type: string
      url:
        type: string
      user_id:
        type: integer
    required:
    - url
    type: object
  controllers.MaterialPageResult:
    properties:
      list:
        items:
          $ref: '#/definitions/models.Material'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  controllers.MaterialUpdateRequest:
    properties:
      title_cn:
        type: string
      title_en:
        type: string
      url:
        type: string
    type: object
...
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  models.Material:
    properties:
      code:
        description: 素材编号
        type: integer
      created_at:
        type: string
      id:
        type: integer
      title_cn:
        description: 中文标题
        type: string
      title_en:
        description: 英文标题
        type: string
      updated_at:
        type: string
      url:
        description: 素材URL
        type: string
      user_id:
        description: 用户ID
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        description: 'admin: 管理员, user: 普通用户'
        type: string
      status:
        description: '1: 正常, 0: 禁用'
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: xhw-service 后台接口文档，业务接口统一返回 utils.Response，鉴权接口请先登录后在 Swagger Authorize
    中填写 Bearer Token。
  title: xhw-service API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，成功后返回用户信息和 JWT Token。
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 用户登录
      tags:
      - 认证
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号，用户名命中管理员白名单时会自动赋予管理员角色。
      parameters:
      - description: 注册参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 用户注册
      tags:
      - 认证
  /api/v1/materials:
    get:
      description: 分页查询素材列表，管理员可按 user_id 筛选任意用户素材，普通用户仅能查询自己的素材。
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，最大100
        in: query
        name: page_size
        type: integer
      - description: 用户ID，管理员可选
        in: query
        name: user_id
        type: integer
      - description: 素材编号
        in: query
        name: code
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.MaterialPageResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取素材列表
      tags:
      - 素材管理
    post:
      consumes:
      - application/json
      description: 创建素材并自动生成当前归属用户下递增的素材编号；普通用户提交的 user_id 会被覆盖为当前登录用户。
      parameters:
      - description: 素材信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.MaterialCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Material'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建素材
      tags:
      - 素材管理
  /api/v1/materials/{id}:
    delete:
      description: 删除素材，普通用户只能删除自己的素材。
      parameters:
      - description: 素材ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除素材
      tags:
      - 素材管理
    get:
      description: 管理员可查看任意素材，普通用户只能查看自己的素材。
      parameters:
      - description: 素材ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Material'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取素材详情
      tags:
      - 素材管理
    put:
      consumes:
      - application/json
      description: 更新素材信息，普通用户只能更新自己的素材，code 和 user_id 不允许修改。
      parameters:
      - description: 素材ID
        in: path
        name: id
        required: true
        type: integer
      - description: 素材更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.MaterialUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Material'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新素材
      tags:
      - 素材管理
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header，格式为 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"